feat(apidemo): show full response body + wrap toggle; full Material Symbols set
apidemo body view:
- Remove the 20000-char cap on the response body (it cut long responses off
  around line ~1100); memoize the pretty-print and syntax highlight so a huge
  body is processed once, not every frame.
- Wrap-aware gutter: line numbers now follow the body's soft-wrap, emitting one
  number per source line with blank rows for wrapped continuations.
- Add a Wrap toggle (response bottom bar): off = no wrap + horizontal scroll
  (one source line per row, gutter pinned), on = soft-wrap.

core/renderers:
- fillMaxWidth/fillMaxHeight are now no-ops under an infinite incoming max
  (matches official Compose). Previously they sized to Int.MAX_VALUE inside a
  scroll, which overflowed the renderers' off-screen cull (x + width) and
  blanked a no-wrap body inside a horizontalScroll.
- Harden the off-screen cull in both renderers with Long math.

material-symbols:
- Generate the full Material Symbols set (4262 icons, exact font codepoints)
  via scripts/generate-material-symbols.py, replacing the curated subset.

Remove TO_FIX.md (the body-truncation bug it tracked is fixed).

+++ renderer-sdl3/src/nativeMain/kotlin/com/compose/desktop/native/renderer/sdl/Sdl3Renderer.kt
@@ -104,8 +104,10 @@ internal class Sdl3Renderer(
         if (vLayer == null || !vLayer.needsTransform) {
             val vTop = inNode.absoluteY
             val vLeft = inNode.absoluteX
-            val vOffscreen = vTop + inNode.height < 0 || vTop > backend.windowHeight ||
-                             vLeft + inNode.width < 0 || vLeft > backend.windowWidth
+            // Long math so a very large node size can't overflow `x + width`
+            // into a negative (which would wrongly read as off-screen).
+            val vOffscreen = vTop.toLong() + inNode.height < 0L || vTop > backend.windowHeight ||
+                             vLeft.toLong() + inNode.width < 0L || vLeft > backend.windowWidth
             if (vOffscreen && (inNode.children.isEmpty() || clipsChildren(inNode))) return
         }
